% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dev_version.R
\name{add_news}
\alias{add_news}
\title{Add a Development Section to \file{NEWS.md}}
\usage{
add_news(path)
}
\arguments{
\item{path}{Path to your package's directory or the \file{NEWS.md} file.}
}
\value{
TRUE on success, FALSE otherwise.
}
\description{
Add a Development Section to \file{NEWS.md}
}
\keyword{internal}
