% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{as.package}
\alias{as.package}
\title{Coerce Input to a Package}
\usage{
as.package(x)
}
\arguments{
\item{x}{object to coerce to a package}
}
\description{
Possible specifications of package:
\itemize{
  \item path
  \item package object
}
}
\keyword{internal}
