% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{author_at_r}
\alias{author_at_r}
\title{Construct a Package Creator String}
\usage{
author_at_r(given, family, email)
}
\arguments{
\item{given}{\code{\link[utils:person]{person}}.}

\item{family}{\code{\link[utils:person]{person}}.}

\item{email}{\code{\link[utils:person]{person}}.}
}
\value{
\code{\link[base:invisible]{Invisibly}}
a list of results of setting the xxx-package.R and the DESCRIPTION.
}
\description{
Construct a Package Creator String
}
\examples{
author_at_r("Your", "Name", "some@whe.re")
}
