% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tests.R
\name{bioc_check}
\alias{bioc_check}
\title{Run BiocCheck on a package}
\usage{
bioc_check(package = getwd(), checkDir)
}
\arguments{
\item{package}{Path to either a package source directory or a package tarball. Default is current working directory.}
}
\value{
An object of class BiocCheck which consists of three lists: error, warning and note.
}
\description{
Run BiocCheck on a package
}
