% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_packages.R
\name{build_and_install_package}
\alias{build_and_install_package}
\title{Build and then optionally install a package}
\usage{
build_and_install_package(
  package_dir = getwd(),
  output_directory = NULL,
  install = TRUE,
  build_vignettes = FALSE,
  bump_version = NULL
)
}
\arguments{
\item{package_dir}{Path to the package source directory. Default is the current directory.}

\item{output_directory}{Directory in which to save package tarball. Default is the parent directory of the package source directory.}

\item{install}{Logical value indicating if package should be installed after it is built. Default is TRUE.}

\item{build_vignettes}{A logical value indicating whether or not to build PDF vignettes. Default is FALSE.}

\item{bump_version}{If provided, bumps the package version using usethis::use_version().
Should be one of "major", "minor", "patch" or "dev".}
}
\value{
Path to the package tarball
}
\description{
First updates documentation and adds any missing dependencies to DESCRIPTION and then builds the package and finally installs it if specified.
}
