% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_manual.R
\name{build_manual}
\alias{build_manual}
\title{Build a Package's Manual}
\usage{
build_manual(
  path = ".",
  output_directory = NULL,
  roxygenise = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{path}{Path to the package.}

\item{output_directory}{Where to put the manual. Defaults to the Rcheck
directory.}

\item{roxygenise}{Run  \pkg{roxygen} first?}

\item{verbose}{Be verbose?}
}
\value{
\code{\link[base:invisible]{Invisibly}} the value of the call to
\code{R CMD Rd2pdf}.
}
\description{
\pkg{devtools} version
(\code{\link[devtools:build_manual]{devtools::build_manual}}) does not run
\pkg{roxygen} first and by defaults puts it in ../ instead of .Rcheck
}
