% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_packages.R
\name{build_vignettes}
\alias{build_vignettes}
\title{Build vignettes for a package}
\usage{
build_vignettes(package_dir = getwd())
}
\arguments{
\item{package_dir}{Path to the package source directory. Default is the current directory.}
}
\description{
Build vignettes for a package
}
