% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shipping.R
\name{calc_shipping}
\alias{calc_shipping}
\title{Calculate Shipping Cost}
\usage{
calc_shipping(weight, zone)
}
\arguments{
\item{weight}{Numeric weight}

\item{zone}{Shipping zone (1–8)}
}
\value{
Shipping cost
}
\description{
Calculate Shipping Cost
}
