% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_cyclometric_complexity}
\alias{check_cyclometric_complexity}
\title{Check Cyclomatic Complexity}
\usage{
check_cyclometric_complexity(path = ".", max_complexity = 10)
}
\arguments{
\item{path}{The package's root directory.}

\item{max_complexity}{The maximum complexity (which should not be exceeded).}
}
\value{
\code{\link[base:invisible]{Invisibly} \link[base:logical]{TRUE}} if
maximum cyclomatic complexity is not exceeded, throws an error otherwise.
}
\description{
Run
\code{\link[cyclocomp:cyclocomp_package_dir]{cyclocomp_package_dir}} on the
package throwing an error when the maximum complexity is exceeded.
}
