% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_news}
\alias{check_news}
\title{Check for \file{NEWS.md} Being Up to Date}
\usage{
check_news(path = ".")
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}
}
\value{
\code{\link[base:invisible]{Invisibly} \link[base:logical]{TRUE}} if
\file{NEWS.md} matches DESCRIPTION, throws an error otherwise.
}
\description{
Compare your \file{NEWS.md} file to the 'Version' entry in DESCRIPTION.
}
\seealso{
Other maintenance functions: 
\code{\link{check_archive}()},
\code{\link{check_codetags}()},
\code{\link{check_cyclomatic_complexity}()},
\code{\link{check_usage}()},
\code{\link{get_check_status}()}
}
\concept{maintenance functions}
