% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chisq_table_context.R
\name{chisq_table_context}
\alias{chisq_table_context}
\title{Test du Chi-2 et analyse contextuelle}
\usage{
chisq_table_context(x)
}
\arguments{
\item{x}{Un tableau de contingence (matrice ou 'table').}
}
\value{
Un objet de classe "chisq_table_context", qui est une liste contenant les résultats du test et les éléments contextuels
}
\description{
Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)
}
\author{
Samira Amani & Samuel Forcier & Taous Sifouane
}
