% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean_orders.R
\name{clean_orders}
\alias{clean_orders}
\title{Clean Orders Data}
\usage{
clean_orders(df)
}
\arguments{
\item{df}{Data frame of orders}
}
\value{
Cleaned data frame
}
\description{
Standardizes order dataset formats.
}
