% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageR_functions.R
\name{create_package}
\alias{create_package}
\title{This function gets your new R-package up and running. 
You only need to work out some clever code.}
\usage{
create_package(wd = "", pname = "", scripts_dir = "", scripts = "",
  install_after = FALSE, replace = TRUE)
}
\arguments{
\item{wd}{Worikng directory (where the package should be created).}

\item{pname}{Package name.}

\item{scripts_dir}{Directory with scripts to be packaged.}

\item{scripts}{Names of scripts to be packaged.}

\item{install_after}{Set to TRUE if the package should be installed after creation.}
}
\description{
This function gets your new R-package up and running. 
You only need to work out some clever code.
}
\examples{
create_package(wd, pname, install=T, scripts_dir, script_name)
}
