% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DESCRIPTION.R
\name{DESCRIPTION}
\alias{DESCRIPTION}
\title{DESCRIPTION}
\usage{
DESCRIPTION(name, date, description, author, email, imports = c("a", "bb"))
}
\arguments{
\item{name}{...}

\item{date}{...}

\item{description}{...}

\item{author}{...}

\item{email}{...}

\item{imports}{... (default : c("a", "bb"))}
}
\value{
...
}
\description{
...
}
\examples{
...
}
