% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equipe4.Chi2-package.R
\docType{package}
\name{Equipe4.Chi2-package}
\alias{Equipe4.Chi2}
\alias{Equipe4.Chi2-package}
\title{Equipe4.Chi2: : Analyse du test du Chi-deux de Pearson avec méthodes S3 (print, summary et plot)}
\description{
Fournit une implémentation complète du test du Chi-deux de Pearson. Le package permet d'appliquer le test sur n'importe quelle table de contingence ou matrice d'effectifs, et génère un objet S3 contenant les valeurs observées et les résidus standardisés de Pearson. Il inclut également des méthodes S3 dédiées (print, summary et plot) pour afficher, résumer et visualiser les résultats du test. Plusieurs types de graphiques sont disponibles, dont une mosaïque, une carte de chaleur des résidus, une carte de chaleur des contributions au chi-deux, des comparaisons observé vs attendu, des proportions par modalité, et un balloon plot. Un jeu de donnérs tiré de HairEyeColor de datasets est inclut dans le package pour illustrer son utilisation.
}
\author{
\strong{Maintainer}: Samira Amani \email{golly-reina-samira.amani.1@ulaval.ca}

Authors:
\itemize{
  \item Samira Samira \email{golly-reina-samira.amani.1@ulaval.ca}
  \item Samuel Forcier \email{samuel.forcier.2@ulaval.ca}
}

}
\keyword{internal}
