% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tests.R
\name{examples_check}
\alias{examples_check}
\title{Check examples in a package}
\usage{
examples_check(package = getwd(), start = NULL)
}
\arguments{
\item{package}{Path to either a package source directory or a package tarball. Default is the current working directory.}

\item{start}{Which file to start running examples from. Default is to start from the first file in lexicographical order.}
}
\description{
Check examples in a package
}
