% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_vignette_code.R
\name{extract_vignette_codes}
\alias{extract_vignette_codes}
\title{Extract and Save R Code Vignettes}
\usage{
extract_vignette_codes(path = ".")
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}
}
\value{
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:NULL]{NULL}}.
}
\description{
I want R code from different kinds of vignettes stored in inst/doc.
}
\keyword{internal}
