% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/files.R
\name{file_length}
\alias{file_length}
\title{Find the number of lines in a file}
\usage{
file_length(path)
}
\arguments{
\item{path}{character; path to file}
}
\value{
numeric; number of lines
}
\description{
Returns the number of lines in a file using the [`readLines()`] function
}
