% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{function_length}
\alias{function_length}
\title{Approximate the length of a function}
\usage{
function_length(body)
}
\arguments{
\item{body}{character; function body as a character vector that has already been passed to the [`language_to_character()`] function}
}
\value{
numeric; length as number
}
\description{
Returns an estimation of the length of a function (in number of lines)
}
