% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{function_list_to_tibble}
\alias{function_list_to_tibble}
\title{Convert a list of functions to a tibble}
\usage{
function_list_to_tibble(function_list, info = FALSE)
}
\arguments{
\item{function_list}{list; a list of the functions. Must contain sublists with a 'name' item}

\item{info}{logical; should the function info (body and length) be included (default is FALSE)? This information must be in the provided list}
}
\value{
tibble; tibble with 1 column ('name') if `info` is FALSE, or 3 columns ('name', 'body' and 'length') if `info` is TRUE
}
\description{
This converts the long function format (a list with names, bodies, etc.) to a tibble. This makes things like plotting much easier.
}
