% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{get_check_status}
\alias{get_check_status}
\title{Retrieve Check Status From a Log File}
\usage{
get_check_status(path)
}
\arguments{
\item{path}{Path to the log file.}
}
\value{
A list :\describe{
    \item{status}{ A list of 
    \describe{
        \item{notes}{The number of \code{NOTE}s}
        \item{warnings}{The number of \code{WARNING}s}
        \item{error}{The number of \code{ERROR}s}
    }}
    \item{log}{A list of 
    \describe{
        \item{notes}{The log entries for \code{NOTE}s}
        \item{warnings}{The log entries for \code{WARNING}s}
        \item{error}{The log entries for \code{ERROR}s}
    }}
}
}
\description{
Searches a typical \command{R CMD check} like \file{xxx.Rcheck/00check.log}
for Errors, Warnings and Notes.
}
\seealso{
Other maintenance functions: 
\code{\link{check_archive}()},
\code{\link{check_codetags}()},
\code{\link{check_cyclomatic_complexity}()},
\code{\link{check_news}()},
\code{\link{check_usage}()}
}
\concept{maintenance functions}
\keyword{internal}
