% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{get_options}
\alias{get_options}
\title{Get Options For Packages}
\usage{
get_options(package_name, ..., remove_names = FALSE, flatten_list = TRUE)
}
\arguments{
\item{package_name}{The package's name.}

\item{...}{See \code{\link{getOption}}.}

\item{remove_names}{[boolean(1)]\cr Remove the names?}

\item{flatten_list}{[boolean(1)]\cr Return a vector?}
}
\value{
A (possibly named) list or a vector.
}
\description{
A convenience function for \code{\link{getOption}}.
}
\seealso{
Other option functions: 
\code{\link{set_options}()}
}
\concept{option functions}
