% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_version.R
\name{get_package_version}
\alias{get_package_version}
\title{Query Installed Package Versions}
\usage{
get_package_version(x, lib_loc = NULL)
}
\arguments{
\item{x}{A character giving the package name.}

\item{lib_loc}{See argument \code{lib.loc} in
\code{\link{packageDescription}}.}
}
\value{
A character giving the package version.
}
\description{
See \code{
\link[fritools:get_package_version]{fritools::get_package_version}
}.
}
