% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_add_commit}
\alias{git_add_commit}
\title{Git Add All Changes and Commit}
\usage{
git_add_commit(
  path,
  message = "Uncommented Changes: Backing Up",
  untracked = FALSE
)
}
\arguments{
\item{path}{The path to the repository.}

\item{message}{The commit message to use.}

\item{untracked}{Add files not tracked yet before committing?}
}
\value{
The return value of
\code{\link[gert:git_commit_all]{gert::git_commit_all}}.
}
\description{
Much like \command{git commit -am"M"}, where \command{M} is the
\code{message}.
}
\seealso{
Other git wrappers: 
\code{\link{git_tag}()}
}
\concept{git wrappers}
