% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_diff}
\alias{git_diff}
\title{Show a Git Diff for a File}
\usage{
git_diff(x, path, verbose = TRUE)
}
\arguments{
\item{x}{The path to the file relative to the repository given by
\code{path}.}

\item{path}{The path to the git repository.}

\item{verbose}{Be verbose? This is the main purpose of this tiny little
wrapper!}
}
\value{
The git diff.
}
\description{
Show a Git Diff for a File
}
