% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git.R
\name{git_tag}
\alias{git_tag}
\title{Create a Git Tag Based on the Current Version Number}
\usage{
git_tag(path = ".", tag_uncommited = FALSE, message = "CRAN release")
}
\arguments{
\item{path}{Path to the package.}

\item{tag_uncommited}{Tag if there are uncommitted changes?}

\item{message}{The tag message to be used.}
}
\value{
\code{\link{FALSE}} or the value of
\code{\link[gert:git_tag_list]{gert::git_tag_list}}.
}
\description{
This is basically the same as \command{git tag -a T -m M} where T is the
version read from the package's DESCRIPTION file and M is given by
\code{message} (see below).
}
\seealso{
Other git wrappers: 
\code{\link{git_add_commit}()}
}
\concept{git wrappers}
