% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Equipe4.Chi2-package.R
\docType{data}
\name{HairEyeColor}
\alias{HairEyeColor}
\title{Couleur des yeux et des cheveux}
\format{
Un \code{data.frame} de 592 observations et 3 variables :
\describe{
\item{\code{Hair}}{Couleur des cheveux (facteur).
Modalités : \code{"Black"}, \code{"Brown"}, \code{"Red"}, \code{"Blond"}.}

\item{\code{Eye}}{Couleur des yeux (facteur).
Modalités : \code{"Brown"}, \code{"Blue"}, \code{"Hazel"}, \code{"Green"}.}

\item{\code{Freq}}{Effectif (nombre d’individus observés). Entier positif.}
}

#' @source
Données originales : \code{datasets::HairEyeColor} de la librarie GLMData
}
\usage{
HairEyeColor
}
\description{
Le jeu de données est la distribution conjointe entre la couleur des cheveux
et celle des yeux d'un groupe d'étudiants en statistique
de la forme \code{Hair × Eye × Sex}.
}
\details{
\itemize{
\item une couleur de cheveux (\code{Hair}),
\item une couleur d’yeux (\code{Eye}),
\item le nombre d’individus observés dans cette combinaison (\code{Freq}).
}
}
\examples{
data(HairEyeColor)

# Table de contingence Hair × Eye
table_contingence <- xtabs(Freq ~ Hair + Eye, data = HairEyeColor)

# Application du test du Chi-deux via ce package
donnees <- chisq_table_context(table_contingence)
summary(donnees)
plot(donnees, type_graphique = "heatmap")
}
\keyword{datasets}
