% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_deps.R
\name{install_deps}
\alias{install_deps}
\title{Install Dependencies From a Package's DESCRIPTION}
\usage{
install_deps(path = ".", verbose = FALSE, ...)
}
\arguments{
\item{path}{Path to the package.}

\item{verbose}{Be verbose?}

\item{...}{passed to
\code{\link[remotes:install_version]{remotes::install_version}}.}
}
\value{
TRUE
}
\description{
\pkg{remotes}' version does not seem to care for the required version. So it
always uses the latest?
We use the lowest version accepted by the definition in file DESCRIPTION.
}
\keyword{internal}
