% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_packages.R
\name{install_package_tarball}
\alias{install_package_tarball}
\title{Install a package from its tarball}
\usage{
install_package_tarball(package_tarball)
}
\arguments{
\item{package_tarball}{Path to a package tarball.}
}
\description{
Install a package from its tarball
}
