% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{is_r_package}
\alias{is_r_package}
\title{Is a Directory an R Package Root Directory?}
\usage{
is_r_package(path)
}
\arguments{
\item{path}{The path to the directory.}
}
\value{
TRUE if the directory is an \R package root directory.
}
\description{
Just a convenience wrapper to
\code{\link[rprojroot:criteria]{rprojroot::is_r_package}}.
}
\keyword{internal}
