% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_version_sufficient.R
\name{is_version_sufficient}
\alias{is_version_sufficient}
\title{Is a Version Requirement Met?}
\usage{
is_version_sufficient(installed, required)
}
\arguments{
\item{required}{The version required.}

\item{available}{The version available.}
}
\value{
\code{\link{TRUE}}, if so,  \code{\link{FALSE}} otherwise.
}
\description{
See \code{
\link[fritools:is_version_sufficient]{fritools::is_version_sufficient}
}.
}
\keyword{internal}
