% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/knn_context_versionR.R
\name{knn_context}
\alias{knn_context}
\title{Contexte pour une classification k-NN avec LLM}
\usage{
knn_context(data, class_var, x1, x2, k = 5, model_ollama = "llama3")
}
\arguments{
\item{data}{data.frame contenant la variable de classe et les prédicteurs.}

\item{class_var}{Nom de la variable de classe (non quoted, ex: Species).}

\item{x1, x2}{Noms des deux variables numériques (non quoted).}

\item{k}{Nombre de voisins k.}

\item{model_ollama}{Nom du modèle Ollama à utiliser.}
}
\value{
Un objet de classe \code{knn_context}.
}
\description{
Cette fonction entraîne un modèle de classification
k plus proches voisins (k-NN) sur deux variables
numériques et une variable de classe, puis demande
à un LLM (Ollama) d'interpréter le résultat.
}
