% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{language_to_character}
\alias{language_to_character}
\title{Convert a language object to a character}
\usage{
language_to_character(lang)
}
\arguments{
\item{lang}{language; language object to be converted}
}
\value{
vector; vector of character strings with roughly one line per entry
}
\description{
Converts a language object (like the body of a function) to a vector of character strings that roughly translate to one line per entry
}
\examples{
language_to_character(functionBody(language_to_character))
}
