% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lintr.R
\name{lint_package}
\alias{lint_package}
\title{Customize \code{\link[lintr:lint_package]{lintr::lint_package}}}
\usage{
lint_package(path)
}
\arguments{
\item{path}{The path to the package, passed to
\code{\link[lintr:lint_package]{lintr::lint_package}}.}
}
\value{
See the return value of
\code{\link[lintr:lint_package]{lintr::lint_package}}.
}
\description{
\pkg{lintr} now runs \pkg{cyclocomp}, which we use indendently and we don't
want to run it twice. So this is just a wrapper to
code{\link[lintr:lint_package]{lintr::lint_package}} where we hardcode the
exclusion of unwanted linters (more may be added to \pkg{lintr}) so other
packages using \pkg{packager}'s \file{Makefile} or
\code{\link{get_package_makelist}} don't have to care of changes to the
default linters in \pkg{lintr}.
}
