% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{list_functions}
\alias{list_functions}
\title{List all functions in a package}
\usage{
list_functions(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See [devtools::as.package()] for more information.}

\item{info}{logical; whether to return additional info (e.g. body and length) for each function. Default is FALSE}

\item{character}{logical; if `info` is TRUE, then should the body of the function be returned as a character string (default) or language object}
}
\value{
list; list of function names only if `info` is FALSE, otherwise a list of names, function bodies and lengths
}
\description{
List all exported functions in a package.
}
