% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ne_utility.R
\name{list_to_function_list}
\alias{list_to_function_list}
\title{Change to format of a list of function names to match that returned from [`function_info()`]}
\usage{
list_to_function_list(list)
}
\description{
Change to format of a list of function names to match that returned from [`function_info()`]
}
\examples{
list_to_function_list(
list = list("sum", "mean")
)
}
