% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{mark_lints}
\alias{mark_lints}
\alias{print_lints}
\title{Mark Lints by Name Suffix}
\usage{
mark_lints(
  x,
  sort = TRUE,
  invert = FALSE,
  file_name_markers = c("_internals", "_verbatim", "_modified")
)

print_lints(
  x,
  sort = TRUE,
  invert = FALSE,
  file_name_markers = c("_internals", "_verbatim", "_modified")
)
}
\arguments{
\item{x}{A list of lints.}

\item{sort}{Sort by file name suffix?}

\item{invert}{Invert the sorting?}

\item{file_name_markers}{Parts of the file name which mark copied code.}
}
\value{
The list of lints with names marked.
}
\description{
I often use internals from other packages and save them in files
named ..._internals..., ..._verbatim... or ..._modified... .
\cr I want these to be marked in \pkg{lintr}'s output.
}
\note{
\code{print_lints} is an old, stale name for \code{mark_lints}.
}
\keyword{internal}
