% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tests.R
\name{package_check_pipeline}
\alias{package_check_pipeline}
\title{Run package checks on a package}
\usage{
package_check_pipeline(
  package_dir = getwd(),
  run_examples_check = TRUE,
  run_r_check = TRUE,
  run_bioc_check = TRUE,
  run_bioc_git_check = TRUE
)
}
\arguments{
\item{package_dir}{Path to a package source directory. Default is current working directory.}

\item{run_examples_check}{Logical value indicating whether to run devtools::examples_check(). Default is TRUE.}

\item{run_r_check}{Logical value indicating whether to run R CMD check. Default is TRUE.}

\item{run_bioc_check}{Logical value indicating whether to run BiocCheck. Default is TRUE.}

\item{run_bioc_git_check}{Logical value indicating whether to run BiocCheckGitClone. Default is TRUE.}
}
\value{
An list with the results from the specified checks.
}
\description{
Run package checks on a package
}
