% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing_DESCRIPTION.R
\name{packing_DESCRIPTION}
\alias{packing_DESCRIPTION}
\title{packing_DESCRIPTION}
\usage{
packing_DESCRIPTION(name, date, description, author, email, imports)
}
\arguments{
\item{name}{...}

\item{date}{...}

\item{description}{...}

\item{author}{...}

\item{email}{...}

\item{imports}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
