% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing_LICENSE.R
\name{packing_LICENSE}
\alias{packing_LICENSE}
\title{packing_LICENSE}
\usage{
packing_LICENSE(name, date, author)
}
\arguments{
\item{name}{...}

\item{date}{...}

\item{author}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
