% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing_Makefile.R
\name{packing_Makefile}
\alias{packing_Makefile}
\title{packing_Makefile}
\usage{
packing_Makefile(git_install)
}
\arguments{
\item{git_install}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
