% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packing_README.R
\name{packing_README}
\alias{packing_README}
\title{packing_README}
\usage{
packing_README(
  name,
  figure_path,
  lifecycle,
  git_install,
  isCovenant,
  description,
  documentation
)
}
\arguments{
\item{name}{...}

\item{figure_path}{...}

\item{lifecycle}{...}

\item{git_install}{...}

\item{isCovenant}{...}

\item{description}{...}

\item{documentation}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
