% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.chisq_table_context}
\alias{plot.chisq_table_context}
\title{Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)}
\usage{
\method{plot}{chisq_table_context}(
  x,
  type_graphique = c("heatmap", "heatmap2", "mosaique", "mosaique2", "barplot",
    "balloon"),
  ...
)
}
\arguments{
\item{x}{Un tableau de contingence (matrice ou 'table').}

\item{type_graphique}{Le type de graphique à afficher (exemple: Heatmap ou mosaique).}

\item{...}{Des arguments additionnels.}
}
\value{
Les résultats du test de khi-2 et le graphique spécifié en entrée
}
\description{
Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)
}
\author{
Samira Amani & Samuel Forcier
}
