% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot_function_length}
\alias{plot_function_length}
\title{Show a boxplot of a package's function lengths}
\usage{
plot_function_length(pkg = ".")
}
\arguments{
\item{pkg}{The package to use, can be a file path to the package or a
package object.  See [devtools::as.package()] for more information.}
}
\value{
ggplot2; ggplot2 graphic (boxplot)
}
\description{
Create a ggplot2 graphic showing the range of function lengths in a package.
}
