% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.chisq_table_context}
\alias{print.chisq_table_context}
\title{Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)}
\usage{
\method{print}{chisq_table_context}(x, ...)
}
\arguments{
\item{x}{Un tableau de contingence (matrice ou 'table').}

\item{...}{Des arguments additionnels.}
}
\value{
Les résultats du test de khi-2 de façon très concise
}
\description{
Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)
}
\author{
Samira Amani & Samuel Forcier
}
