% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{print_lints}
\alias{print_lints}
\title{Print Lints by Name Suffix}
\usage{
print_lints(x, sort = TRUE, invert = FALSE,
  file_name_markers = c("_internals", "_verbatim", "_modified"))
}
\arguments{
\item{x}{A list of lints.}

\item{sort}{Sort by file name suffix?}

\item{invert}{Invert the sorting?}

\item{file_name_markers}{Parts of the file name which mark copied code.}
}
\value{
The list of lints with names marked.
}
\description{
I often use internals from other packages and save them in files 
named ..._internals..., ..._verbatim... of ..._modified... .
\cr I want these to be marked in lintr's output.
}
\examples{
files <- list.files(system.file("files", package = "packager"), 
                    full.names = TRUE)
lints <- lintr:::flatten_lints(lapply(files, function(file) {
                                          if (interactive()) {
                                              message(".", appendLF = FALSE)
                                          }
                                          lintr::lint(file, 
                                                      parse_settings = FALSE)
                                     }))

print_lints(lints, invert = FALSE)
print_lints(lints, invert = TRUE)
}
