% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/various.R
\name{provide_make}
\alias{provide_make}
\title{Provide File \command{make.R}}
\usage{
provide_make(path, Rbuildignore = TRUE, ...)
}
\arguments{
\item{path}{Where to create the file.}

\item{Rbuildignore}{Add the file to \file{.Rbuildignore} under the given
\code{path}?}

\item{...}{Arguments passed to
\code{\link[packager:use_template]{packager::use_template}}.}
}
\value{
\code{\link[=invisible]{Invisibly}}
\code{\link[=logical]{TRUE}} on success,
\code{\link[=invisible]{Invisibly}}
\code{\link[=logical]{FALSE}} otherwise.
}
\description{
Provide File \command{make.R}
}
\keyword{internal}
