% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/makelist.R
\name{provide_make_list}
\alias{provide_make_list}
\title{Provide a Makelist Suitable for Packages with \pkg{packager}.}
\usage{
provide_make_list()
}
\description{
Provide a Makelist Suitable for Packages with \pkg{packager}.
}
