% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/news_rd.R
\name{provide_news_rd}
\alias{provide_news_rd}
\title{Convert \file{NEWS.md} to \file{inst/NEWS.rd}}
\usage{
provide_news_rd(path = ".", force = is_force(), verbose = TRUE)
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}

\item{force}{Overwrite an existing file?}

\item{verbose}{Be verbose?}
}
\value{
\code{\link{TRUE}} on success, \code{\link{FALSE}} otherwise.
}
\description{
The file \file{inst/NEWS.rd} will get listed in the package's help index.
Christian Vonderach uses the index thus wants \file{inst/NEWS.rd} to
automatically be in sync with \file{./NEWS.md}.
}
\keyword{internal}
