% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package_tests.R
\name{r_check}
\alias{r_check}
\title{Run R CMD check on a package}
\usage{
r_check(package)
}
\arguments{
\item{package}{Path to either a package source directory or a package tarball. Default is the current working directory.}
}
\value{
An object of class rcmdcheck.
}
\description{
Run R CMD check on a package
}
