% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/README.R
\name{README.md}
\alias{README.md}
\title{README.md}
\usage{
README.md(
  name,
  figure_path,
  lifecycle,
  git_install,
  isCovenant,
  description,
  documentation
)
}
\arguments{
\item{name}{...}

\item{figure_path}{...}

\item{lifecycle}{...}

\item{git_install}{...}

\item{isCovenant}{...}

\item{description}{...}

\item{documentation}{...}
}
\value{
...
}
\description{
...
}
\examples{
...
}
