% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/release.R
\name{release}
\alias{release}
\title{Release a Package to CRAN}
\usage{
release(path = ".", stop_on_git = TRUE, force = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{The package's root directory.}

\item{stop_on_git}{Stop if git has uncommitted changes or is not synced with
the origin?}

\item{force}{Skip user interaction?}

\item{verbose}{Be verbose?}
}
\value{
\code{\link[base:invisible]{Invisibly} \link{NULL}}
}
\description{
This is a stripped version of \pkg{devtools}'
\code{\link[devtools:release]{release function}}, it does not do all the
interactive checks.
}
