% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_packages.R
\name{reload_package}
\alias{reload_package}
\title{Detach and then reload a package}
\usage{
reload_package(package)
}
\arguments{
\item{package}{A string giving the name of a package}
}
\description{
Detach and then reload a package
}
