% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/options.R
\name{set_options}
\alias{set_options}
\title{Set Options For Packages}
\usage{
set_options(package_name, ..., overwrite = TRUE)
}
\arguments{
\item{package_name}{The package's name.}

\item{...}{See \code{\link{options}}.}

\item{overwrite}{[boolean(1)]\cr Overwrite options already set?}
}
\value{
invisible(TRUE)
}
\description{
See \code{\link[fritools:set_options]{fritools:set_options}}
}
\seealso{
Other option functions: 
\code{\link{get_options}()}
}
\concept{option functions}
\keyword{internal}
