% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sort_deps_in_desc.R
\name{sort_deps_in_desc}
\alias{sort_deps_in_desc}
\title{Sort Fields `Import`, `Depends` and `Suggests`  \file{DESCRIPTION}}
\usage{
sort_deps_in_desc(file)
}
\arguments{
\item{file}{see
\code{\link[desc:desc_get_deps]{desc::desc_get_deps}}).}
}
\value{
\code{\link{NULL}}, but called for its side effect.
}
\description{
Sort Fields `Import`, `Depends` and `Suggests`  \file{DESCRIPTION}
}
\examples{
path <- file.path(tempdir(), "myFirstPackage")
usethis::create_package(path = path, rstudio = FALSE, open = FALSE)
withr::with_dir(path, usethis::use_package("withr"))
withr::with_dir(path, usethis::use_package("cleanr"))
desc::desc_get_deps(file.path(path, "DESCRIPTION"))
packager::sort_deps_in_desc(file.path(path, "DESCRIPTION"))
usethis::proj_set(NULL)
desc::desc_get_deps(file.path(path, "DESCRIPTION"))
unlink(path, recursive = TRUE)
}
\keyword{internal}
