% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageR.R
\name{source_dev}
\alias{source_dev}
\title{source_dev}
\usage{
source_dev(name = "packageR", dev_path = ".")
}
\arguments{
\item{name}{... (default : "packageR")}

\item{dev_path}{... (default : ".")}
}
\value{
...
}
\description{
...
}
\examples{
...
}
