% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{submit}
\alias{submit}
\alias{release}
\title{Release a Package to CRAN}
\usage{
submit(
  path = ".",
  stop_on_git = TRUE,
  stop_on_devel = TRUE,
  force = FALSE,
  verbose = TRUE
)

release(
  path = ".",
  stop_on_git = TRUE,
  stop_on_devel = TRUE,
  force = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{path}{The package's root directory.}

\item{stop_on_git}{Stop if git has uncommitted changes or is not synced with
the origin?}

\item{stop_on_devel}{Stop if the package has a developement version? (That is, a four part version.)}

\item{force}{Skip user interaction?}

\item{verbose}{Be verbose?}
}
\value{
\code{\link[base:invisible]{Invisibly} \link{NULL}}
}
\description{
This is a stripped version of \pkg{devtools}'
\code{\link[devtools:release]{release function}}, omitting most of the
interactive checks.
}
\note{
\code{release} is just a link to \code{submit} as
\code{\link[devtools:release]{release}} is the original function from
\pkg{devtools}.
}
