% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary.chisq_table_context}
\alias{summary.chisq_table_context}
\title{Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)}
\usage{
\method{summary}{chisq_table_context}(object, ...)
}
\arguments{
\item{object}{Un tableau de contingence (matrice ou 'table').}

\item{...}{Des arguments additionnels.}
}
\value{
Les résultats du test de khi-2 de façon un peu plus détaillée
}
\description{
Effectue le test du Chi-2 sur un tableau de contingence et prépare un objet
pour l'analyse S3 pour impression (print), résumé (summary) et graphique (plot)
}
\author{
Samira Amani & Samuel Forcier & Taous Sifouane
}
