% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_deps.R
\name{update_deps}
\alias{update_deps}
\title{Update Dependencies for a Package Directory}
\usage{
update_deps(path)
}
\arguments{
\item{path}{the package's root directory.}
}
\description{
\pkg{cyclocomp} updates the dependencies using \pkg{remotes}. Well, for some
strange reason it does so on every run (temporarily). So I permanently do
this, and \code{\link[remotes:update_packages]{remotes::update_packages}}
does so only for CRAN packages, and \code{remotes:::update.package_deps} is
internal to \pkg{remotes} only. So I need copies of internal functions from
\pkg{remotes}.
}
\keyword{internal}
