% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageR_functions.R
\name{update_package}
\alias{update_package}
\title{This function updates a package.}
\usage{
update_package(wd = "", pname = "", install_after = FALSE)
}
\arguments{
\item{wd}{Worikng directory (where the package repository is located).}

\item{pname}{Package name.}
}
\description{
This function updates a package.
}
\examples{
update_package(wd, pname, install=T)
}
