% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_verbatim.R
\name{use_build_ignore}
\alias{use_build_ignore}
\title{Add files to \file{.Rbuildignore}}
\usage{
use_build_ignore(files, escape = TRUE, pkg = ".")
}
\arguments{
\item{files}{Paths of files.}

\item{escape}{If \code{TRUE}, the default, will escape \code{.} to
\code{\\.} and surround with \code{^} and \code{$}.}

\item{pkg}{Path to the package directory (see
\code{\link[devtools:as.package]{as.package}}).}
}
\value{
Nothing, called for its side effect.
}
\description{
This is verbatim copy of  git commit
\verb{a5e5805ecd630ebc46e080bd78ebcf32322efe3c} of \pkg{usethis}.\cr
\file{.Rbuildignore} has a regular expression on each line, but it's
usually easier to work with specific file names. By default, will (crudely)
turn filenames into  regular expressions that will only match these
paths. Repeated entries will be silently removed.
}
