% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dev_version.R
\name{use_dev_news}
\alias{use_dev_news}
\title{Use a Development Version in NEWS.md}
\usage{
use_dev_news(path = ".")
}
\arguments{
\item{path}{Path to your package's directory or the the NEWS file.}
}
\value{
TRUE on success, FALSE otherwise.
}
\description{
Use a Development Version in NEWS.md
}
