% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools.R
\name{use_directory}
\alias{use_directory}
\title{Use a Directory}
\usage{
use_directory(path, ignore = FALSE, pkg = ".")
}
\arguments{
\item{path}{Path of the directory to create, relative to the project.}

\item{ignore}{Add the directory to \file{.Rbuildignore}?}

\item{pkg}{Path to the package directory (see
\code{\link[devtools:as.package]{as.package}}).}
}
\description{
Create a directory.\cr
this is verbatim copy of  git commit
\verb{a5e5805ecd630ebc46e080bd78ebcf32322efe3c} of \pkg{usethis}.
}
