% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{use_git_check_version_not_tagged}
\alias{use_git_check_version_not_tagged}
\title{Check to Not Commit to a Tagged Version}
\usage{
use_git_check_version_not_tagged(path)
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}
}
\value{
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:NULL]{NULL}}.
}
\description{
I sometimes forget to bump the package version of a tagged state in git, 
then commiting git to that version. Nothing servere, but should be avoided.
So we add a check (to) a pre-commit hook for git.
}
