% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/git_hooks.R
\name{use_git_pre_commit_script}
\alias{use_git_pre_commit_script}
\title{Use a Script as git pre-commit hook}
\usage{
use_git_pre_commit_script(path, script_file)
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}

\item{script_file}{A path to a script file.}
}
\description{
Provide a pre-commit.d/ in ./git/hooks, a pre-commit hook that executes
scripts from that directory and copy the \emph{script_file} there.
}
\details{
This might work only for proper git, not for packages like \pkg{gert} or
\pkg{git2r}.
}
\seealso{
Other infrastructure functions: 
\code{\link{use_git_check_version_not_tagged}()}
}
\concept{infrastructure functions}
\keyword{internal}
