% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{use_template}
\alias{use_template}
\title{Use a Template}
\usage{
use_template(
  template,
  save_as = template,
  data = list(),
  ignore = FALSE,
  pkg = ".",
  source_package = "packager",
  git_commit = TRUE,
  force = isTRUE(getOption("packager")[["force"]])
)
}
\arguments{
\item{template}{The name of the packages's template (a file in the template
directory).}

\item{save_as}{The path where the template should be written to.}

\item{ignore}{Add the new file to \file{.Rbuildignore}?}

\item{pkg}{The path to the package template shall be used for.}

\item{source_package}{The name of the package from which the template should
be read.}

\item{git_commit}{Add the new file to git and commit?}

\item{force}{Overwrite an existing file?}
}
\value{
\code{\link{TRUE}} on success, \code{\link{FALSE}} otherwise.
}
\description{
\pkg{devtools}' or, later \pkg{usethis}' function hardcoded the source
package, the had dialogs on opened files in RStudio. Did Stuff I do not want.
}
\keyword{internal}
