% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/upload_data_zenodo.R
\name{zenodo_upload}
\alias{zenodo_upload}
\title{Upload a file to Zenodo}
\usage{
zenodo_upload(file, zenodo_access_token, deposit_id)
}
\arguments{
\item{file}{Path to file to upload.}

\item{zenodo_access_token}{Path to Zenodo access token. Tokens can be created at
https://zenodo.org/account/settings/applications/tokens/new/ with the "deposit:write" and "deposit:actions" scopes.}

\item{deposit_id}{The deposit ID of the upload. This should be the last part of the URL for the deposit.}
}
\value{
Returns TRUE or FALSE depending on if the file was successfully uploaded.
}
\description{
Upload a file to Zenodo
}
