% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_and_install_packages.R
\name{add_dependencies}
\alias{add_dependencies}
\title{Add dependencies to DESCRIPTION file of a package and optionally install them if they are not already installed}
\usage{
add_dependencies()
}
\value{
A vector with the names of dependencies.
}
\description{
Add dependencies to DESCRIPTION file of a package and optionally install them if they are not already installed
}
