% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{add_github_url_to_desc}
\alias{add_github_url_to_desc}
\title{Add a github URL to File DESCRIPTION}
\usage{
add_github_url_to_desc(path = ".", default_gh_user = NULL,
  normalize = TRUE)
}
\arguments{
\item{path}{The path to the package.}

\item{path}{default_gh_user See details.}
}
