% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/devtools_verbatim.R
\name{as.package}
\alias{as.package}
\title{Coerce input to a package.}
\usage{
as.package(x = NULL, create = NA)
}
\arguments{
\item{x}{object to coerce to a package}

\item{create}{only relevant if a package structure does not exist yet: if
`TRUE`, create a package structure; if `NA`, ask the user
(in interactive mode only)}
}
\description{
Possible specifications of package:
\itemize{
  \item path
  \item package object
}
}
\keyword{internal}
