% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_codetags}
\alias{check_codetags}
\title{Check for Code Tags}
\usage{
check_codetags(path = ".", exclude_pattern = "\\\\.Rcheck/",
  include_pattern = "\\\\.[Rr]$|\\\\.[Rr]md$",
  pattern = "XXX:|FIXME:|TODO:")
}
\arguments{
\item{path}{The directory to search.}

\item{exclude_pattern}{A pattern for exlusions based on the file names.
Stronger than \code{include_pattern}.}

\item{include_pattern}{A pattern for inclusions based on the file names.}

\item{pattern}{The pattern to search for.}
}
\value{
A character vector of hits.
}
\description{
You hopefully use code tags
(see \href{PEP 350}{https://www.python.org/dev/peps/pep-0350/} for example).
This functions searches for files under a directory containing such tags.
}
\examples{
dir <- system.file("templates", package = "packager")
check_codetags(dir)
}
