% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{check_news}
\alias{check_news}
\title{Check for NEWS.md Being Up to Date}
\usage{
check_news(path = ".")
}
\arguments{
\item{path}{The directory to search.}
}
\value{
\code{\link[base:invisible]{Invisibly} \link[base:logical]{TRUE}} if 
NEWS.md matches DESCRIPTION, throws an error otherwise.
}
\description{
Compare your NEWS.md file to the 'Version' entry in DESCRIPTION.
}
