% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checks.R
\name{check_usage}
\alias{check_usage}
\title{Check Usage with \pkg{codetools}' \code{\link{checkUsagePackage}}}
\usage{
check_usage(path = ".")
}
\arguments{
\item{path}{Path to the package directory.}
}
\value{
A character vector of issues found by
\code{\link{checkUsagePackage}}.
}
\description{
This is just a convenience wrapper to \code{\link{checkUsagePackage}} (which
needs loading of the [development version of the] package).
}
