% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create}
\alias{create}
\title{Create a Package Template}
\usage{
create(path, force = TRUE, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{structure}}.}

\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}
}
\value{
FALSE. But it does not return anything, it stops with a
condition of class c("error", "packager", "condition").
}
\description{
This is just a wrapper to create a package using
\code{\link[devtools:create]{devtools::create}} and
infect it using \code{\link{infect}}.
}
\examples{
tryCatch(packager:::throw("Hello error!"), packager = function(e) return(e))
}
