% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{create}
\alias{create}
\title{Create a Package Template}
\usage{
create(path, force = TRUE, ...)
}
\arguments{
\item{path}{See \code{\link[devtools:create]{devtools::create}}.}

\item{force}{Recursively \code{\link{unlink}} the path before calling
\code{\link[devtools:create]{devtools::create}(path)}?}

\item{...}{Arguments to be passed to \code{\link{structure}}.}
}
\value{
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:NULL]{NULL}}.
condition of class c("error", "packager", "condition").
}
\description{
This is just a wrapper to create a package using
\code{\link[devtools:create]{devtools::create}} and
infect it using \code{\link{infect}}.
}
\examples{
path <- file.path(tempdir(), "myPackage")
packager::create(path = path)
list.files(path)
unlink(path, recursive = TRUE)
}
