% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gitlab.R
\name{get_gitlab_log}
\alias{get_gitlab_log}
\title{Read a Gitlab Check Log}
\usage{
get_gitlab_log(user, project, private_token, ...)
}
\arguments{
\item{user}{The user's name on gitlab.}

\item{project}{The project's name on gitlab.}

\item{private_token}{The user's private token on gitlab.}

\item{...}{Arguments passed to \code{\link[httr:GET]{httr::GET}}.}
}
\value{
A character vector containing the lines of the gitlab log.
}
\description{
For a given user's project, the last log for jobs for name and stage "check"
will be read. This is assumed to be the output of R CMD check,
rcmdcheck::rcmdcheck(), devtools::check() or the like.
}
\examples{
\dontrun{
if (Sys.info()[["nodename"]] == "fvafrdebianCU") {
    j <- get_gitlab_log(user = "fvafrcu", project = "packager", 
                        private_token = readLines(file.path("~", 
                                                  ".gitlab_private_token.txt")), 
                        httr::use_proxy("10.127.255.17", 8080))
} else {
    j <- get_gitlab_log(user = "fvafrcu", project = "packager", 
                        private_token = readLines(file.path("~", 
                                                  ".gitlab_private_token.txt"))) 
}

cat(j, sep = "\\n")
}
}
