% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internal.R
\name{git_add_commit}
\alias{git_add_commit}
\title{Git Add All Changes and Commit}
\usage{
git_add_commit(repo, message = "Uncommented Changes: Backing Up")
}
\arguments{
\item{repo}{The repository to commit.}

\item{message}{The commit message to use.}
}
\value{
The return value of \code{\link{gitr::commit}}.
}
\description{
The same as git commit -am"M", where M is the \code{message}.
}
