% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{infect}
\alias{infect}
\title{Throw a Condition}
\usage{
infect(path, make = TRUE, git_add_and_commit = TRUE, ...)
}
\arguments{
\item{...}{Arguments to be passed to \code{\link{structure}}.}

\item{message_string}{The message to be thrown.}

\item{system_call}{The call to be thrown.}
}
\value{
FALSE. But it does not return anything, it stops with a
condition of class c("error", "packager", "condition").
}
\description{
Throws a condition of class c("error", "packager", "condition").
}
\details{
We use this condition as an error dedicated to \pkg{ packager }.
}
\examples{
tryCatch(packager:::throw("Hello error!"), packager = function(e) return(e))
}
\keyword{internal}
