% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{infect}
\alias{infect}
\title{Adjust a Package}
\usage{
infect(path, make = FALSE, git_add_and_commit = TRUE, ...)
}
\arguments{
\item{path}{Path to the package to get infected.}

\item{make}{Try to use make on the infected package?
Option to be disabled for testing.}

\item{git_add_and_commit}{Add and commit changes in git? 
Option to be disabled for testing.}

\item{...}{Arguments to be passed to \code{\link{set_package_info}}.}
}
\value{
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:NULL]{NULL}}.
condition of class c("error", "packager", "condition").
}
\description{
Add a variety of extensions to a package (skeleton).
}
\details{
This is \pkg{packager}'s core.
}
\examples{
path <- file.path(tempdir(), "myPackage")
devtools::create(path = path)
l1 <- list.files(path, recursive = TRUE, full.names = TRUE)
packager::infect(path = path)
unlink(path, recursive = TRUE)
l2 <- list.files(path, recursive = TRUE, full.names = TRUE)
print(l1); print(l2)
}
