% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packager-package.R
\docType{package}
\name{packager-package}
\alias{packager-package}
\title{Helps Me Build Packages}
\description{
Helper functions for a build system, heavily borrowing from and
extending package \pkg{devtools} 1.13.3.
}
\details{
I find \pkg{devtools} very helpful but it lacks some functionality I would
want
while creating and building a package.  Maybe \pkg{packager} should be two
packages.
}
\keyword{package}
