% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packager-package.R
\docType{package}
\name{packager-package}
\alias{packager-package}
\title{Helps Me Build Packages}
\description{
This is a set of wrappers around \pkg{devtools} and \pkg{MakefileR} and some
sanity checks for developing packages.
}
\keyword{package}
