% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageR.R
\name{packing}
\alias{packing}
\title{packing}
\usage{
packing(
  package_path = ".",
  Rfile_pattern = ".*[.]R$",
  name = "packageR",
  date = Sys.Date(),
  description = "A simple R package that helps create simple R packages.",
  documentation =
    "Navigate to a directory with R files, open an R console, and run `?packageR::packing` to see available arguments. Then, run `packageR::packing(*arguments*)` to create your R package.",
  author = "Louis Héraut",
  email = "louis.heraut@inrae.fr",
  imports = c("roxygen2", "remotes", "devtools", "stringr"),
  git_install = "remotes::install_github('super-lou/packageR')",
  figure_path = NULL,
  lifecycle = "Stable",
  add_file = c("CODE_OF_CONDUCT.md", "DESCRIPTION", "LICENSE", "Makefile", "README.md")
)
}
\arguments{
\item{package_path}{Path to the directory to be transformed into an R package (default: ".")}

\item{Rfile_pattern}{Regular expression pattern to select specific files in the package_path directory (default: ".*\link{.}R$")}

\item{name}{Name of the package (default: "packageR")}

\item{date}{Creation date of the package or version (default: Sys.Date())}

\item{description}{Package description (default: "A simple R package that helps create simple R packages.")}

\item{documentation}{Simple README documentation explaining how the package works (default: "Navigate to a directory with R files, open an R console, and run \code{?packageR::packing} to see available arguments. Then, run \verb{packageR::packing(*arguments*)} to create your R package.")}

\item{author}{Package author (default: "Louis Héraut")}

\item{email}{Author's email (default: "louis.heraut@inrae.fr")}

\item{imports}{Required packages for package installation (default: c("roxygen2", "remotes", "devtools", "stringr"))}

\item{git_install}{R command to install packages from a specified repository (default: "remotes::install_github('super-lou/packageR')")}

\item{figure_path}{Path to a figure to use in the README}

\item{lifecycle}{Project lifecycle status (default: "Stable")
Lifecycle statuses:
\itemize{
\item Experimental: The project is in its early stages of development, and the codebase may change frequently.
\item Maturing: The codebase is being refined, but finer details are likely to change.
\item Stable: The project is in a reliable state, and major changes are unlikely to occur.
\item Dormant: The project is not currently under active development, but there are plans for future redevelopment.
\item Retired: The project is no longer in use and/or supported.
}}

\item{add_file}{List of additional files commonly included in packages (default: c("CODE_OF_CONDUCT.md", "DESCRIPTION", "LICENSE", "Makefile", "README.md"))}
}
\value{
A new directory in the working directory formatted as a usable R package, based on the provided input files.
}
\description{
This function creates an R package from a selection of R files containing functions.
}
\examples{
# for default option (which is the option used for create this R package)
packageR::packing()

# for your personal package
packageR::packing(package_path="/home/alice/CheshiRCat/",
                  Rfile_pattern=".*[.]R$",
                  name="CheshiRCat",
                  date="1865-11-01",
                  description="Twas brillig, and the slithy toves. Did gyre and gimble in the wabe; All mimsy were the borogoves, And the mome raths outgrabe.",
                  documentation="Beware the [*Jabberwock*](https://en.wikipedia.org/wiki/Jabberwocky), my son! The jaws that bite, the claws that catch!</br> Beware the *Jubjub bird*, and shun The frumious *Bandersnatch*!</br> He took his **vorpal sword** in hand:</br></br> - Long time the manxome foe he sought </br> - So rested he by the Tumtum tree, And stood awhile in thought.</br></br> And as in uffish thought he stood `?CheshiRCat::smile`",
                  author="Alice",
                  email="alice@wonderland.uk",
                  imports=c("remotes", "stringr"),
                  git_install="remotes::install_github('alice_adventures/CheshiRCat')",
                  figure_path=NULL,
                  lifecycle="Dormant",
                  add_file=c("CODE_OF_CONDUCT.md", "DESCRIPTION", "LICENSE", "Makefile", "README.md"))
}
