% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/packageR.R
\name{packing}
\alias{packing}
\title{packing}
\usage{
packing(
  Rfile_pattern = ".*[.]R$",
  name = "packageR",
  date = Sys.Date(),
  description = "A simple R package that help to create simple R package.",
  documentation =
    "Go into a directory where there are some R files and within a R console you can type `?packageR::packing` to see possible arguments and then type `packageR::packing(*arguments*)` to create your R package.",
  author = "Louis Héraut",
  email = "louis.heraut@inrae.fr",
  imports = c("roxygen2", "remotes", "devtools"),
  git_install = "remotes::install_github('super-lou/packageR')",
  figure_path = "AE.png",
  lifecycle = "Stable",
  add_file = c("CODE_OF_CONDUCT.md", "DESCRIPTION", "LICENSE", "Makefile", "README.md")
)
}
\arguments{
\item{Rfile_pattern}{... (default : ".*[.]R$")}

\item{name}{... (default : "packageR")}

\item{date}{... (default : Sys.Date())}

\item{description}{... (default : "A simple R package that help to create simple R package.")}

\item{documentation}{... (default : "Go into a directory where there are some R files and within a R console you can type `?packageR::packing` to see possible arguments and then type `packageR::packing(*arguments*)` to create your R package.")}

\item{author}{... (default : "Louis Héraut")}

\item{email}{... (default : "louis.heraut@inrae.fr")}

\item{imports}{... (default : c("roxygen2", "remotes", "devtools"))}

\item{git_install}{... (default : "remotes::install_github('super-lou/packageR')")}

\item{figure_path}{... (default : "AE.png")}

\item{lifecycle}{... (default : "Stable")}

\item{add_file}{... (default : c("CODE_OF_CONDUCT.md", "DESCRIPTION", "LICENSE", "Makefile", "README.md"))}
}
\value{
...

...

...
}
\description{
...

...

...
}
\examples{
...
...
...
}
