% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_comments.R
\name{provide_cran_comments}
\alias{provide_cran_comments}
\title{Provide a Template for Your Comments To CRAN}
\usage{
provide_cran_comments(check_log = NULL, path = ".", initial = FALSE,
  travis_session_info = NULL, write_to_file = TRUE,
  private_token = NULL, name = NA, proxy = NULL)
}
\arguments{
\item{check_log}{Path to the check log relative to \code{path}. Typically
file.path("log", "check.Rout").}

\item{path}{The path to the package.}

\item{initial}{Is this an initial release?}

\item{travis_session_info}{\url{https://travis.com} session info, search
for\cr
\code{$ Rscript -e 'sessionInfo()'} \cr
in the raw log of the \url{https://travis.com} build and copy the
following three lines. This could read\cr
\preformatted{
travis_session_info <- c("
                    R version 3.4.0 (2017-04-21)
                    Platform: x86_64-pc-linux-gnu (64-bit)
                    Running under: Ubuntu precise (12.04.5 LTS)
                    ")}
Or provide a path to a file containing the current \verb{travis} log.\cr
Set to \dQuote{\verb{travis-cli}} to retrieve Session info automatically if
your system is
set up to use \url{https://github.com/travis-ci/travis.rb}.}

\item{write_to_file}{Do write the comment to \file{cran-comment.md}?}

\item{private_token}{A private token to access \url{https://gitlab.com}.}

\item{name}{The name to sign with, if NA, the given name of the package
maintainer as stated in file DESCRIPTION is used.}

\item{proxy}{A proxy to use.}
}
\value{
Character vector containing the \acronym{CRAN} comments, which are
written to \file{cran-comments.md} (see Note).
}
\description{
\code{\link{release}} reads a file \file{cran-comments.md}. This
function provides a template based on your R version your
\command{R CMD check log} and
the package's \file{NEWS.md}.
}
\note{
By default this function writes to disk as side effect.
}
\examples{
\dontrun{
gitlab_token <- NULL
gitlab_token <- readLines(file.path("~", ".gitlab_private_token.txt"))

check_log <- system.file("files", "check.Rout", package = "packager")
tsi <- system.file("files", "travis.log",
                                   package = "packager")
if (Sys.info()[["nodename"]] == "fvafrdebianCU") {
    proxy <- httr::use_proxy("10.127.255.17", 8080)
    comments <- provide_cran_comments(path = ".",
                                      check_log = check_log,
                                      travis_session_info = tsi,
                                      write_to_file = TRUE,
                                      private_token = gitlab_token,
                                      proxy = proxy)

} else {
    comments <- provide_cran_comments(path = ".",
                                      check_log = check_log,
                                      travis_session_info = tsi,
                                      write_to_file = FALSE,
                                      private_token = gitlab_token)
}
cat(comments, sep = "")
}
}
