% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cran_comments.R
\name{provide_cran_comments}
\alias{provide_cran_comments}
\title{Provide a Template For Your Comments To CRAN}
\usage{
provide_cran_comments(check_log, path = ".", travis_raw_log = NULL)
}
\arguments{
\item{check_log}{Path to the check log relative to \code{path}. Typically 
file.path("log", "dev_check.Rout").}

\item{path}{The path to the package.}

\item{travis_raw_log}{Travis check info, search for\cr
"$ Rscript -e 'sessionInfo()'" \cr 
in the raw log of the travis build and copy the
following three lines. This could read\cr 
travis_raw_log <- c("\cr 
                    R version 3.4.0 (2017-04-21)\cr 
                    Platform: x86_64-pc-linux-gnu (64-bit)\cr 
                    Running under: Ubuntu precise (12.04.5 LTS) \cr 
                    ")\cr}
}
\value{
NULL.
}
\description{
Devtools' \code{\link{release}} reads a file \emph{cran-comments.md}. This
function provides a template based on your R version and your check log.
}
