% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{provide_gitlab_url}
\alias{provide_gitlab_url}
\title{Provide a \code{gitlab} \code{url} for a Given Path}
\usage{
provide_gitlab_url(path = ".")
}
\arguments{
\item{path}{The path to the directory or one of its subdirectories.}
}
\value{
a character string giving a github \code{url}.
}
\description{
Provide a \code{gitlab} \code{url} for a Given Path
}
\examples{
path <- file.path(tempdir(), "foo")
unlink(path, recursive = TRUE)
devtools::create(path)
try(provide_gitlab_url(path))
git2r::init(path)
provide_gitlab_url(path)
invisible(desc::desc_set(Package = "bar", file = path))
provide_gitlab_url(path)
}
