% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{provide_make}
\alias{provide_make}
\title{Provide File \command{make.R}}
\usage{
provide_make(path, Rbuildignore = TRUE)
}
\arguments{
\item{path}{Where to create the file.}

\item{Rbuildignore}{Add the file to .Rbuildignore under the given
\code{path}?}
}
\value{
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:logical]{TRUE}} on success, 
\code{\link[base:invisible]{Invisibly}}
\code{\link[base:logical]{FALSE}} otherwise.
}
\description{
Provide File \command{make.R}
}
