% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rcmdcheck.R
\name{rcmdcheck_and_log}
\alias{rcmdcheck_and_log}
\title{Run \code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck}} and Write to Log}
\usage{
rcmdcheck_and_log(
  path = ".",
  args = c(build_args, "--as-cran"),
  build_args = NULL
)
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}

\item{args}{Arguments passed to
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck}}.}

\item{build_args}{Arguments passed to
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck}}.}
}
\value{
\code{\link[=invisible]{Invisibly}  \link{NULL}}.
}
\description{
The \code{\link{deparse}}d  \code{\link{dput}}s of \code{\link{Sys.info}} and
\code{\link[rcmdcheck:rcmdcheck]{rcmdcheck::rcmdcheck}} are tagged and
\code{\link{cat}}ed so we can evaluate them from reading logs (on
\verb{gitlab}, for example).
}
\keyword{internal}
