% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tools.R
\name{set_desc_url}
\alias{set_desc_url}
\title{Set a DESCRIPTION File's \code{URL} Field}
\usage{
set_desc_url(url, path = ".", normalize = TRUE, overwrite = FALSE,
  do_commit = is_force(),
  do_remind = !isTRUE(getOption("packager")[["force"]]),
  verbose = getOption("packager")[["verbose"]])
}
\arguments{
\item{url}{A character string giving the \code{URL} to set or add in
DESCRIPTION.}

\item{path}{Path to the DESCRIPTION file, see
\code{\link[desc:desc_get_urls]{desc::desc_get_urls}}.}

\item{normalize}{See \code{\link[desc:desc_set_urls]{desc::desc_set_urls}}.}

\item{overwrite}{Set (overwrite) the \code{URL} field in DESCRIPTION instead
adding the \code{URL} given to the \code{URL} field in DESCRIPTION?}

\item{do_commit}{Commit the updated DESCRIPTION?}

\item{do_remind}{Write a reminder into the package's TODO.md?}

\item{verbose}{Be verbose?}
}
\value{
\code{\link[base:invisible]{Invisibly} \link[base:logical]{TRUE}}
}
\description{
I frequently forget to add an \code{URL} to my packages' DESCRIPTION files,
and when I do not, I often forget to check that the \code{URL} is valid,
respectively the one I want. \cr
So this is a wrapper to functions from \pkg{desc} and \pkg{git2r} and i
messaging and/or adding
a reminder to file \code{TODO.md}.
}
\examples{
path <- file.path(tempdir(), "foo")
unlink(path, recursive = TRUE)
devtools::create(path)
repo <- git2r::init(path)
git2r::add(repo = repo, path = "*")
git2r::commit(repo = repo, message = "Initial commit")
url <- provide_gitlab_url(path = path)
set_desc_url(url, path = path)
git2r::commits(repo)
git2r::status(repo)
readLines(file.path(path, "TODO.md"))
}
