% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{set_package_info}
\alias{set_package_info}
\title{Set a Package's Info}
\usage{
set_package_info(path, author_at_r = NULL,
  title = "What it Does (One Line, Title Case)", description = NULL,
  details = NA)
}
\arguments{
\item{path}{Path to the package.}

\item{author_at_r}{A string giving the author. See \code{\link{author_at_r}}.}

\item{title}{A string giving the title.}

\item{description}{A string giving the description.}

\item{details}{A string giving the details. Defaults to NA in which case a
default details are inserted. Set to NULL to have no details at all.}
}
\value{
\code{\link[base:invisible]{Invisibly}}
a list of results of setting the xxx-package.R and the DESCRIPTION.
}
\description{
Fill DESCRIPTION and R/xxx-package.R with the same Title and Description,
keeping the info given in both places identical.
}
\examples{
path <- file.path(tempdir(), "myPackage")
devtools::create(path = path)
a  <- sub("(email)", "\\n\\t\\\\1",
          packager::author_at_r("Your", "Name", "some@whe.re"))
set_package_info(path = path, author_at_r = a, title = "What Now?",
                 description = "This package does nothing.",
                 details = "Details do not show up in DESCRIPTION.")
package_desc <- file.path(path, "DESCRIPTION")
package_info_file <- file.path(path,
                               "R", paste0(basename(path), "-package.R"))
readLines(package_desc)
readLines(package_info_file)
unlink(path, recursive = TRUE)
}
