% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dev_version.R
\name{use_dev_version}
\alias{use_dev_version}
\title{Use a Development Version in DESCRIPTION and \file{NEWS.md}}
\usage{
use_dev_version(path = ".", force = FALSE)
}
\arguments{
\item{path}{Path to the package directory (see
\code{\link[devtools:as.package]{devtools::as.package}}).}

\item{force}{Set to \code{TRUE} to force version bumping with uncommitted git
changes.}
}
\description{
This is much like
\code{\link[usethis:use_dev_version]{usethis::use_dev_version}}, but their
conventions keep changing.
}
