% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dev_version.R
\name{use_dev_version}
\alias{use_dev_version}
\title{Use a Development Version in DESCRPITION and NEWS.md}
\usage{
use_dev_version(path)
}
\arguments{
\item{path}{Path to your package's directory.}
}
\value{
The return value of \code{\link{use_dev_news}}.
}
\description{
Use a Development Version in DESCRPITION and NEWS.md
}
